+++ stream-chat-android-ui-common/src/main/kotlin/com/getstream/sdk/chat/viewmodel/messages/MessageListViewModel.kt
@@ -30,6 +30,7 @@ import io.getstream.chat.android.livedata.ChatDomain
 import io.getstream.chat.android.livedata.controller.ChannelController
 import io.getstream.chat.android.offline.experimental.channel.state.MessagesState
 import io.getstream.chat.android.offline.experimental.extensions.asReferenced
+import io.getstream.chat.android.offline.extensions.downloadAttachment
 import io.getstream.chat.android.offline.extensions.setMessageForReply
 import kotlinx.coroutines.flow.map
 import kotlin.properties.Delegates
@@ -324,7 +325,7 @@ public class MessageListViewModel @JvmOverloads constructor(
                 client.setMessageForReply(event.cid, event.repliedMessage).enqueue()
             }
             is Event.DownloadAttachment -> {
-                domain.downloadAttachment(event.attachment).enqueue(
+                client.downloadAttachment(event.attachment).enqueue(
                     onError = { chatError ->
                         logger.logE("Attachment download error: ${chatError.message}. Cause: ${chatError.cause?.message}")
                     }